+++ src/TitleDetails/BoundItemsList/utils.js
@@ -1,8 +1,8 @@
 import { getItemStatusLabel } from '@folio/stripes-acq-components';
 import { NoValue, TextLink } from '@folio/stripes/components';
 
-import { PIECE_COLUMNS } from '../constants';
 import { getPieceStatusFromItem } from '../../common/utils';
+import { PIECE_COLUMNS } from '../constants';
 
 export const getColumnFormatter = (hasViewInventoryPermissions, instanceId) => {
   return ({