+++ src/TitleUnreceive/TitleUnreceiveContainer.js
@@ -14,13 +14,11 @@ import {
 import {
   baseManifest,
   batchFetch,
-  itemsResource,
   LIMIT_MAX,
   LINES_API,
   locationsManifest,
   PIECE_STATUS,
   piecesResource,
-  requestsResource,
   useShowCallout,
 } from '@folio/stripes-acq-components';
 
@@ -38,6 +36,10 @@ import {
   CENTRAL_RECEIVING_ROUTE,
   RECEIVING_ROUTE,
 } from '../constants';
+import {
+  usePieceItemsFetch,
+  usePieceRequestsFetch,
+} from '../common/hooks/usePaginatedPieces/hooks';
 import { useReceivingSearchContext } from '../contexts';
 import TitleUnreceive from './TitleUnreceive';
 
@@ -56,7 +58,12 @@ function TitleUnreceiveContainer({
   const [pieceHoldingMap, setPieceHoldingMap] = useState();
   const poLineId = title?.poLineId;
 
-  const { isCentralRouting } = useReceivingSearchContext();
+  const { crossTenant, isCentralRouting, targetTenantId } = useReceivingSearchContext();
+  const { fetchPieceItems } = usePieceItemsFetch({
+    instanceId: title?.instanceId,
+    tenantId: targetTenantId,
+  });
+  const { fetchPieceRequests } = usePieceRequestsFetch({ tenantId: targetTenantId });
 
   useEffect(
     () => {
@@ -90,7 +97,16 @@ function TitleUnreceiveContainer({
             query: `${filterQuery} sortby locationId`,
           },
         })
-          .then(piecesResponse => getHydratedPieces(piecesResponse, mutator.requests, mutator.items))
+          .then(async (piecesResponse) => {
+            const hydratedPieces = await getHydratedPieces({
+              pieces: piecesResponse,
+              fetchPieceItems,
+              fetchPieceRequests,
+              crossTenant,
+            });
+
+            return hydratedPieces;
+          })
           .then(hydratedPieces => {
             setPieces(hydratedPieces);
             const holdingIds = hydratedPieces.map(({ holdingId }) => holdingId).filter(Boolean);
@@ -186,10 +202,6 @@ TitleUnreceiveContainer.manifest = Object.freeze({
     fetch: false,
     tenant: '!{tenantId}',
   },
-
-  // TODO: fetch items and requests (after MODORDERS-1138) from related tenants
-  items: itemsResource,
-  requests: requestsResource,
   locations: {
     ...locationsManifest,
     fetch: false,