+++ src/common/hooks/usePaginatedPieces/hooks/usePieceRequestsFetch.js
@@ -8,7 +8,6 @@ export const usePieceRequestsFetch = ({ tenantId } = {}) => {
   const fetchPieceRequests = ({ pieces, signal }) => {
     const kyExtended = ky.extend({ signal });
 
-    // TODO: integrate loading of requests from several tenants after implementation MODORDERS-1138
     return fetchLocalPieceRequests(kyExtended, { pieces });
   };