+++ src/common/hooks/useQuickReceive.test.js
@@ -18,6 +18,7 @@ import {
 
 jest.mock('../utils', () => ({
   ...jest.requireActual('../utils'),
+  extendKyWithTenant: jest.fn(),
   getItemById: jest.fn(() => () => Promise.resolve({})),
   getPieceById: jest.fn(() => () => Promise.resolve({ json: () => ({}) })),
 }));